move decompressor test
